Refactoring react_pizza
