+++ src/components/SortPopup.jsx
@@ -13,7 +13,7 @@ function SortPopup({ items }) {
   const handleOutsideClick = (e) => {
     if (!e.path.includes(sortRef.current)) {
       setVisiblePopup(false);
-      console.log('outside');
+      /* console.log('outside'); */
     }
   };
